requirements.txt
Babel==2.5.3
Django==2.0.4
django-crispy-forms==1.7.2
Jinja2==2.10
MarkupSafe==1.0
psycopg2==2.7.4
pytz==2018.3
